+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Please mark all change in change log and use the issue from GitHub
 -   \#2692 Milvus hangs during multi-thread concurrent search
 -   \#2739 Fix mishards start failed
 -   \#2767 fix a bug of getting wrong nprobe limitation in knowhere on GPU version
+-   \#2774 Server down during loading data
 -   \#2776 Fix too many data copies during creating IVF index
 
 ## Feature